+++ numpy/setup.py
@@ -17,6 +17,7 @@ def configuration(parent_package='',top_path=None):
     config.add_subpackage('polynomial')
     config.add_subpackage('random')
     config.add_subpackage('testing')
+    config.add_subpackage('typing')
     config.add_data_dir('doc')
     config.add_data_files('py.typed')
     config.add_data_files('*.pyi')
