+++ numpy/typing/__init__.py
@@ -1,0 +1,3 @@
+from ._array_like import _SupportsArray, ArrayLike
+from ._shape import _Shape, _ShapeLike
+from ._dtype_like import DtypeLike
